% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/proverbs.R
\name{proverb}
\alias{proverb}
\title{Print a Welcome to Night Vale proverb to the console}
\usage{
proverb(id)

proverb(id)
}
\arguments{
\item{id}{}
}
\value{



}
\description{
Print a Welcome to Night Vale proverb to the console

Title
}
