% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nv_data.R
\docType{data}
\name{proverb_data}
\alias{proverb_data}
\title{WTNV dataset}
\format{
A data frame:
\describe{
  \item{id}{Unique ID}
  \item{number}{Episode number}
  \item{title}{Episode title}
  \item{airdate}{Episode airdate}
  \item{proverb}{Proverb text}
  }
}
\source{
{Extracted from the WTNV Fandom site. Available at
  https://nightvale.fandom.com/wiki/List_of_Proverbs}
}
\usage{
proverb_data
}
\description{
WTNV dataset
}
\keyword{datasets}
