% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nv_proverb.R
\name{proverb}
\alias{proverb}
\title{Print a Welcome to Night Vale (WTNV) proverb to the console}
\usage{
proverb(id, show_meta = TRUE, show_id = FALSE, as_string = FALSE)
}
\arguments{
\item{id}{If no ID is specified, then a random WTNV proverb is returned.}

\item{show_meta}{Whether to display the metadata (episode title and airdate)
for the proverb. Default is TRUE.}

\item{show_id}{Whether to display the ID of the proverb. Default is FALSE.}

\item{as_string}{If FALSE (default), print proverb to the console. If TRUE,
return proverb as string (useful if, for example, you want to pass the
proverb to the cowsay package).}
}
\description{
prin
}
